# Don't use a regex in IsImplicitGlob