--- a/internal/tsoptions/tsconfigparsing.go
+++ b/internal/tsoptions/tsconfigparsing.go
@@ -105,9 +105,12 @@ func (c *configFileSpecs) matchesInclude(fileName string, comparePathsOptions ts
 		return false
 	}
 	for _, spec := range c.validatedIncludeSpecs {
-		matcher := vfs.NewGlobMatcher(spec, comparePathsOptions.CurrentDirectory, comparePathsOptions.UseCaseSensitiveFileNames)
-		if matcher.MatchesFile(fileName) {
-			return true
+		includePattern := vfs.GetPatternFromSpec(spec, comparePathsOptions.CurrentDirectory, "files")
+		if includePattern != "" {
+			includeRegex := vfs.GetRegexFromPattern(includePattern, comparePathsOptions.UseCaseSensitiveFileNames)
+			if match, err := includeRegex.MatchString(fileName); err == nil && match {
+				return true
+			}
 		}
 	}
 	return false