… internal/vfs/utilities.go
… internal/vfs/utilities_test.go