This code is a hotspot in the critical loading path. In Strada, this was all regexes, which is great in JS where regexes are fast. In Go, not so much, even moreso when we use `regexp2` in ECMAScript compatibility mode. We've always wanted to rewrite this, it's just taken a bit.

This PR moves the old implementation to a different file, then I had agent mode write tests and a new implementation, verifying against those tests.

The result so far:

```
                                          │   old.txt    │               new.txt               │
                                          │    sec/op    │   sec/op     vs base                │
MatchFiles/CommonPattern-20                 186.58µ ± 1%   25.47µ ± 2%  -86.35% (p=0.000 n=10)
MatchFiles/SimpleInclude-20                  77.13µ ± 1%   11.87µ ± 1%  -84.61% (p=0.000 n=10)
MatchFiles/EmptyIncludes-20                  29.00µ ± 2%   11.16µ ± 1%  -61.52% (p=0.000 n=10)
MatchFiles/HiddenDirectories-20              55.31µ ± 1%   13.39µ ± 1%  -75.79% (p=0.000 n=10)
MatchFiles/NodeModulesSearch-20             219.55µ ± 1%   24.57µ ± 1%  -88.81% (p=0.000 n=10)
MatchFiles/LargeFileSystem-20               23.402m ± 1%   2.043m ± 1%  -91.27% (p=0.000 n=10)
MatchFilesLarge/AllFiles-20                 20.920m ± 1%   1.831m ± 1%  -91.25% (p=0.000 n=10)
MatchFilesLarge/Components-20               2308.1µ ± 1%   677.5µ ± 1%  -70.65% (p=0.000 n=10)
MatchFilesLarge/TestFiles-20                2763.0µ ± 1%   338.4µ ± 1%  -87.75% (p=0.000 n=10)
MatchFilesLarge/NestedUtilsWithPattern-20   1212.2µ ± 2%   363.5µ ± 1%  -70.01% (p=0.000 n=10)
geomean                                      673.4µ        113.2µ       -83.19%

                                          │    old.txt    │                new.txt                 │
                                          │     B/op      │     B/op       vs base                 │
MatchFiles/CommonPattern-20                  15.83Ki ± 0%    11.05Ki ± 0%   -30.16% (p=0.000 n=10)
MatchFiles/SimpleInclude-20                  6.797Ki ± 0%    8.117Ki ± 0%   +19.42% (p=0.000 n=10)
MatchFiles/EmptyIncludes-20                  3.802Ki ± 0%    8.031Ki ± 0%  +111.22% (p=0.000 n=10)
MatchFiles/HiddenDirectories-20             10.752Ki ± 0%    8.633Ki ± 0%   -19.71% (p=0.000 n=10)
MatchFiles/NodeModulesSearch-20              20.54Ki ± 0%    15.44Ki ± 0%   -24.83% (p=0.000 n=10)
MatchFiles/LargeFileSystem-20                823.3Ki ± 0%   1190.6Ki ± 0%   +44.62% (p=0.000 n=10)
MatchFilesLarge/AllFiles-20                  751.4Ki ± 0%   1139.3Ki ± 0%   +51.62% (p=0.000 n=10)
MatchFilesLarge/Components-20                243.4Ki ± 0%    532.2Ki ± 0%  +118.67% (p=0.000 n=10)
MatchFilesLarge/TestFiles-20                 117.7Ki ± 0%    263.8Ki ± 0%  +124.14% (p=0.000 n=10)
MatchFilesLarge/NestedUtilsWithPattern-20    127.4Ki ± 0%    290.0Ki ± 0%  +127.66% (p=0.000 n=10)
geomean                                      53.82Ki         74.55Ki        +38.52%

                                          │   old.txt   │                new.txt                │
                                          │  allocs/op  │  allocs/op    vs base                 │
MatchFiles/CommonPattern-20                  335.0 ± 0%     244.0 ± 0%   -27.16% (p=0.000 n=10)
MatchFiles/SimpleInclude-20                  143.0 ± 0%     151.0 ± 0%    +5.59% (p=0.000 n=10)
MatchFiles/EmptyIncludes-20                  86.00 ± 0%    145.00 ± 0%   +68.60% (p=0.000 n=10)
MatchFiles/HiddenDirectories-20              237.0 ± 0%     216.0 ± 0%    -8.86% (p=0.000 n=10)
MatchFiles/NodeModulesSearch-20              333.0 ± 0%     290.0 ± 0%   -12.91% (p=0.000 n=10)
MatchFiles/LargeFileSystem-20               8.829k ± 0%   14.286k ± 0%   +61.81% (p=0.000 n=10)
MatchFilesLarge/AllFiles-20                 8.190k ± 0%   14.208k ± 0%   +73.48% (p=0.000 n=10)
MatchFilesLarge/Components-20               3.108k ± 0%    7.044k ± 0%  +126.64% (p=0.000 n=10)
MatchFilesLarge/TestFiles-20                1.604k ± 0%    3.546k ± 0%  +121.07% (p=0.000 n=10)
MatchFilesLarge/NestedUtilsWithPattern-20   1.661k ± 0%    3.626k ± 0%  +118.30% (p=0.000 n=10)
geomean                                      849.0         1.199k        +41.24%
```

So, somewhere in the 6x to 10x faster range?